+++ spec/ExternalURILoader.spec.ts
@@ -1,9 +1,0 @@
-import ExternalURILoader from "ExternalURILoader";
-
-describe("ExternalURILoader", () => {
-  const loader = new ExternalURILoader();
-
-  it("has executeRequest method", () => {
-    expect(loader.executeRequest).toBeDefined();
-  });
-});