+++ types/GeocoderProviderFactory.d.ts
@@ -1,6 +1,6 @@
-import { BingProvider, GoogleAPIProvider, MapboxProvider, MapboxProviderOptionsInterface, MapquestProvider, NominatimProvider, NominatimProviderOptionsInterface, YandexProvider, YandexProviderOptionsInterface, ProviderOptionsInterface } from "./providers";
+import { BingProvider, GoogleAPIProvider, MapboxProvider, MapboxProviderOptionsInterface, MapquestProvider, NominatimProvider, NominatimProviderOptionsInterface, OpenCageProvider, OpenCageProviderOptionsInterface, YandexProvider, YandexProviderOptionsInterface, ProviderOptionsInterface } from "./provider";
 interface ProviderOptionInterface {
-    provider: "bing" | "google" | "mapbox" | "mapquest" | "nominatim" | "openstreetmap" | "yandex";
+    provider: "bing" | "google" | "mapbox" | "mapquest" | "nominatim" | "opencage" | "openstreetmap" | "yandex";
 }
 interface ProviderFactoryOptions extends ProviderOptionsInterface, ProviderOptionInterface {
 }
@@ -10,12 +10,15 @@ interface MapboxGeocoderProviderFactoryOptions extends ProviderOptionInterface, 
 interface NominatimGeocoderProviderFactoryOptions extends ProviderOptionInterface, NominatimProviderOptionsInterface {
     provider: "nominatim" | "openstreetmap";
 }
+interface OpenCageGeocoderProviderFactoryOptions extends ProviderOptionInterface, OpenCageProviderOptionsInterface {
+    provider: "opencage";
+}
 interface YandexGeocoderProviderFactoryOptions extends ProviderOptionInterface, YandexProviderOptionsInterface {
     provider: "yandex";
 }
-export declare type GeocoderProviderFactoryOptions = ProviderFactoryOptions | MapboxGeocoderProviderFactoryOptions | NominatimGeocoderProviderFactoryOptions | YandexGeocoderProviderFactoryOptions;
-export declare type GeocoderProvider = BingProvider | GoogleAPIProvider | MapboxProvider | MapquestProvider | NominatimProvider | YandexProvider;
-export declare type GeocoderProviderByOptionsType<O> = O extends MapboxGeocoderProviderFactoryOptions ? MapboxProvider : O extends NominatimGeocoderProviderFactoryOptions ? NominatimProvider : O extends YandexGeocoderProviderFactoryOptions ? YandexProvider : GeocoderProvider;
+export declare type GeocoderProviderFactoryOptions = ProviderFactoryOptions | MapboxGeocoderProviderFactoryOptions | NominatimGeocoderProviderFactoryOptions | OpenCageGeocoderProviderFactoryOptions | YandexGeocoderProviderFactoryOptions;
+export declare type GeocoderProvider = BingProvider | GoogleAPIProvider | MapboxProvider | MapquestProvider | NominatimProvider | OpenCageProvider | YandexProvider;
+export declare type GeocoderProviderByOptionsType<O> = O extends MapboxGeocoderProviderFactoryOptions ? MapboxProvider : O extends NominatimGeocoderProviderFactoryOptions ? NominatimProvider : O extends OpenCageGeocoderProviderFactoryOptions ? OpenCageProvider : O extends YandexGeocoderProviderFactoryOptions ? YandexProvider : GeocoderProvider;
 export default class ProviderFactory {
     /**
      * Creates Geocoder Provider instances.