+++ types/error/index.d.ts
@@ -1,0 +1,2 @@
+export { default as ResponseError } from "./ResponseError";
+//# sourceMappingURL=index.d.ts.map