+++ types/providers/mapbox/index.d.ts
@@ -1,7 +1,0 @@
-export { default as MapboxGeocodeQuery } from "./MapboxGeocodeQuery";
-export * from "./MapboxGeocodeQuery";
-export { default as MapboxProvider } from "./MapboxProvider";
-export * from "./MapboxProvider";
-export { default as MapboxReverseQuery } from "./MapboxReverseQuery";
-export * from "./MapboxReverseQuery";
-//# sourceMappingURL=index.d.ts.map